+++ test/languages/csg/test_transform.py
@@ -4,7 +4,7 @@ import torch
 from mlprogram import Environment
 from mlprogram.languages.csg import Interpreter, show
 from mlprogram.languages.csg.ast import Rectangle
-from mlprogram.languages.csg.transform import AddTestCases, TransformCanvas
+from mlprogram.languages.csg.transforms import AddTestCases, TransformCanvas
 
 
 class TestAddTestCases(object):
