+++ test_integration/test_csg_by_pbe_with_repl.py
@@ -29,7 +29,7 @@ from mlprogram.languages.csg import (
     Parser,
     get_samples,
 )
-from mlprogram.languages.csg.transform import AddTestCases, TransformCanvas
+from mlprogram.languages.csg.transforms import AddTestCases, TransformCanvas
 from mlprogram.nn import MLP, AggregatedLoss, Apply, CNN2d
 from mlprogram.nn.action_sequence import Loss
 from mlprogram.nn.pbe_with_repl import Encoder
